+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- Allow to lazy load children nodes
+- Allow to lazy load explorer items to improve performance
+- Add setting to configure to if empty repositories should be listed or not
 
 ### Fixed